+++ src/Facility.Core/Http/HttpContentSerializer.cs
@@ -8,13 +8,7 @@ public abstract class HttpContentSerializer
 	/// <summary>
 	/// Creates a standard HTTP content serializer.
 	/// </summary>
-	/// <remarks>Values are serialized and deserialized via memory streams so that the actual I/O can be asynchronous
-	/// whether or not the serialization format supports asynchronous I/O directly. This also makes it simple to calculate
-	/// the length of the content. Consider using <c>Microsoft.IO.RecyclableMemoryStream</c> to improve performance
-	/// by setting <c>memoryStreamCreator</c> to <c>RecyclableMemoryStreamManager.GetStream</c>. Otherwise
-	/// <c>System.IO.MemoryStream</c> is used.</remarks>
-	public static HttpContentSerializer Create(ServiceSerializer serviceSerializer, Func<Stream>? memoryStreamCreator = null) =>
-		new StandardHttpContentSerializer(serviceSerializer, memoryStreamCreator);
+	public static HttpContentSerializer Create(ServiceSerializer serviceSerializer) => new StandardHttpContentSerializer(serviceSerializer);
 
 	/// <summary>
 	/// The default media type for the serializer.
@@ -86,4 +80,6 @@ public abstract class HttpContentSerializer
 	/// Reads a DTO from the specified HTTP content.
 	/// </summary>
 	protected abstract Task<ServiceResult<object>> ReadHttpContentAsyncCore(Type objectType, HttpContent content, CancellationToken cancellationToken);
+
+	internal static HttpContentSerializer Legacy { get; } = Create(NewtonsoftJsonServiceSerializer.Instance);
 }