+++ src/Facility.Core/Http/JsonHttpContentSerializer.cs
@@ -1,3 +1,7 @@
+using System.Net;
+using System.Net.Http.Headers;
+using Newtonsoft.Json;
+
 namespace Facility.Core.Http;
 
 /// <summary>
@@ -24,9 +28,8 @@ public class JsonHttpContentSerializer : HttpContentSerializer
 	/// </summary>
 	public JsonHttpContentSerializer(JsonHttpContentSerializerSettings? settings)
 	{
-		// we always force async I/O now, so we can ignore settings.ForceAsyncIO
+		m_forceAsyncIO = settings?.ForceAsyncIO ?? false;
 		m_memoryStreamCreator = settings?.MemoryStreamCreator;
-		m_contentSerializer = Create(JsonServiceSerializer.Legacy, CreateMemoryStream);
 
 		SupportedMediaTypes = new[] { HttpServiceUtility.JsonMediaType };
 	}
@@ -54,15 +57,94 @@ public class JsonHttpContentSerializer : HttpContentSerializer
 	/// <summary>
 	/// Creates HTTP content for the specified DTO.
 	/// </summary>
-	protected override HttpContent CreateHttpContentCore(object content, string? mediaType) =>
-		m_contentSerializer.CreateHttpContent(content, mediaType ?? DefaultMediaType);
+	protected override HttpContent CreateHttpContentCore(object content, string? mediaType)
+	{
+		var memoryStream = CreateMemoryStream();
+		ServiceJsonUtility.ToJsonStream(content, memoryStream);
+		return new DelegateHttpContent(mediaType ?? DefaultMediaType, memoryStream);
+	}
 
 	/// <summary>
 	/// Reads a DTO from the specified HTTP content.
 	/// </summary>
-	protected override Task<ServiceResult<object>> ReadHttpContentAsyncCore(Type objectType, HttpContent content, CancellationToken cancellationToken) =>
-		m_contentSerializer.ReadHttpContentAsync(objectType, content, cancellationToken);
+	protected override async Task<ServiceResult<object>> ReadHttpContentAsyncCore(Type objectType, HttpContent content, CancellationToken cancellationToken)
+	{
+		try
+		{
+			if (m_forceAsyncIO)
+			{
+				// read content into memory so that ASP.NET Core doesn't complain about synchronous I/O during JSON deserialization
+				using var stream = CreateMemoryStream();
+#if NET6_0_OR_GREATER
+				await content.CopyToAsync(stream, cancellationToken).ConfigureAwait(false);
+#else
+				await content.CopyToAsync(stream).ConfigureAwait(false);
+#endif
+				stream.Seek(0, SeekOrigin.Begin);
+				return ReadJsonStream(objectType, stream);
+			}
+			else
+			{
+#if NET6_0_OR_GREATER
+				using var stream = await content.ReadAsStreamAsync(cancellationToken).ConfigureAwait(false);
+#else
+				using var stream = await content.ReadAsStreamAsync().ConfigureAwait(false);
+#endif
+				return ReadJsonStream(objectType, stream);
+			}
+		}
+		catch (JsonException exception)
+		{
+			return ServiceResult.Failure(HttpServiceErrors.CreateInvalidContent(exception.Message));
+		}
+	}
 
+	private static ServiceResult<object> ReadJsonStream(Type objectType, Stream stream)
+	{
+		using var textReader = new StreamReader(stream);
+		var deserializedContent = ServiceJsonUtility.FromJsonTextReader(textReader, objectType);
+		if (deserializedContent is null)
+			return ServiceResult.Failure(HttpServiceErrors.CreateInvalidContent("Content must not be empty."));
+		return ServiceResult.Success(deserializedContent);
+	}
+
+	private sealed class DelegateHttpContent : HttpContent
+	{
+		public DelegateHttpContent(string mediaType, Stream memoryStream)
+		{
+			Headers.ContentType = MediaTypeHeaderValue.Parse(mediaType);
+
+			m_memoryStream = memoryStream;
+		}
+
+		protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context)
+		{
+			m_memoryStream.Position = 0;
+			return m_memoryStream.CopyToAsync(stream);
+		}
+
+		protected override bool TryComputeLength(out long length)
+		{
+			length = m_memoryStream.Length;
+			return true;
+		}
+
+		protected override void Dispose(bool disposing)
+		{
+			try
+			{
+				if (disposing)
+					m_memoryStream.Dispose();
+			}
+			finally
+			{
+				base.Dispose(disposing);
+			}
+		}
+
+		private readonly Stream m_memoryStream;
+	}
+
+	private readonly bool m_forceAsyncIO;
 	private readonly Func<Stream>? m_memoryStreamCreator;
-	private readonly HttpContentSerializer m_contentSerializer;
 }