+++ tests/Facility.ConformanceApi.UnitTests/ServiceSerializerTestsBase.cs
@@ -1,5 +1,4 @@
 using Facility.Core;
-using Microsoft.IO;
 
 namespace Facility.ConformanceApi.UnitTests;
 
@@ -21,6 +20,4 @@ public abstract class ServiceSerializerTestsBase
 	protected ServiceSerializer Serializer { get; }
 
 	protected JsonServiceSerializer JsonSerializer { get; }
-
-	protected static RecyclableMemoryStreamManager MemoryStreamManager { get; } = new();
 }